♻️ MapModal과 hooks에서 작성된 useMap 훅을 하나로 사용하도록 리팩토링
